pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "toml"]
build-backend = "setuptools.build_meta"

[project]
name = "studentclearinghouse"
version = "0.1.3"
dependencies = [
    "pandas",
    "pyyaml",
]
description = "A module for working with National Student Clearinghouse submission or return files."
readme = "README.md"
authors = [{ name = "David Onder", email = "donder@haywood.edu" }]
license = {file = "LICENSE"}